requirements.txt
argon2-cffi==23.1.0
fastapi==0.108.0
loguru==0.7.2
pydantic==2.5.3
pydantic-settings==2.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.9
tortoise-orm[accel,asyncpg]==0.20.0
uvicorn[standard]==0.25.0
